make SaltpackSenderIdentify understand revoked keys
There's also a security fix here. Previously we were taking the server's
word that the user reported by key/basics.json actually owned the key we
were asking about. Now we check it ourselves. The key difference is that
now we get all of our final results from LoadUser, which walks the
sigchain and checks the Merkle tree, and which might in some distant
future cross-check the Merkle tree against the blockchain.